PROG-179 Added TODOs for tomorrow, getattr fix pending test
