+++ voidvision/pipeline.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 class VisionPipeline(ABC):
 
 	@abstractmethod
-	def __init__(self, config: str, camera: int, cam_name: str, output_name: str) -> None:
+	def __init__(self, config: str, cam_num: int, cam_name: str, output_name: str, table) -> None:
 		pass
 	
 	@abstractmethod
